Moved inner FileHealthCheckBuilder class to its own class
